Staging

See merge request cse1105/2019-2020/repos/oopp-group-39/repository-template!42

+++ server/src/main/java/nl/tudelft/oopp/group39/reservable/repositories/ReservableRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReservableRepository extends JpaRepository<Reservable, Integer> {
+public interface ReservableRepository extends JpaRepository<Reservable, Long> {
 }
